Fix workflow config for building non-tagged commits
`git describe` does not work for non-tagged commits after fetching
because the commit we're on is still shallow, i.e. it has no history to
it. Use `git fetch --shallow-exclude` instead to unshallow parts of the
history.

Also mount the complete git repository in the build container instead of
the submodule only, otherwise `git describe` is not even executed in the
kernel's `setlocalversion` script, because the `git rev-parse
--show-cdup` check in there fails if the container has no access to the
parent repository.

And finally pass UID and GID from the parent system to the docker
container, otherwise `git rev-parse --show-cdup` in the kernel's
`setlocalversion` complains about dubious ownership of the git repo.